% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As_parm.r
\name{As_parm}
\alias{As_parm}
\title{As_parm Function
This function calculates the As-parameter from Table 1 of Tobiason and O'Melia (1988). The As-parameter is based on the p-parameter.}
\usage{
As_parm(POR = 0.21)
}
\arguments{
\item{POR}{porosity. Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}
}
\description{
As_parm Function
This function calculates the As-parameter from Table 1 of Tobiason and O'Melia (1988). The As-parameter is based on the p-parameter.
}
\examples{
As_parm(POR = 0.3)
[1] 75.49263

As_parm(POR = 21)
[1] "Error: Porosity value must be less than 1"
}
\keyword{As-parameter}
\keyword{filtration,}
\keyword{porosity,}

