% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc0_filt.r
\name{cc0_filt}
\alias{cc0_filt}
\title{cc0_filt Function
This function calculates the ratio of TSS concentration to original TSS concentration at a distance
of LENGTH. The default units are "m" but distance units (LENGTH_UNIT) of cm and ft are also applicable.}
\usage{
cc0_filt(LENGTH = 500, LENGTH_UNIT = "m", alpha = 1, Kx = 0.02,
  GRAD = 7e-04, POR = 0.21, K_units = "cm/s", TEMP = 25,
  DIA_SUSP = 1e-04, DIA_SOIL = 0.053, PART_DENS = 2.64)
}
\arguments{
\item{LENGTH}{Distance at which to calculate the ratio of TSS concentration to original TSS concentration.
Defaults to a value of 500}

\item{LENGTH_UNIT}{Unit of length measurement. Defaults to a value of "m". Other options are "meter", 
"meters", "cm", "centimeter", "centimeters", "ft", "foot", "feet".}

\item{alpha}{Factor that accounts for chemical affects on particle removal. Defaults to a value of 1 (i.e., 
chemical affects do not enhance particle removal}

\item{Kx}{Defaults to a value of 0.02 cm/s}

\item{GRAD}{Defaults to a value of 0.0007}

\item{POR}{Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}

\item{K_units}{Units of hydraulic conductivity. Defaults to cm/s but also accepts ft/day. Units should be surrounded
by quotation marks to ensure it is treated as a string of text.}

\item{TEMP}{Temperature in units of °C. Defaults to a value of 25°C}

\item{DIA_SUSP}{Diameter of the suspended solids in units of cm}

\item{DIA_SOIL}{Diameter of the soil particles in units of cm}

\item{PART_DENS}{Density of the suspended particle in units of g/cm³}
}
\description{
cc0_filt Function
This function calculates the ratio of TSS concentration to original TSS concentration at a distance
of LENGTH. The default units are "m" but distance units (LENGTH_UNIT) of cm and ft are also applicable.
}
\examples{
L_filt()
[1] 0.2625636
}
\keyword{concentration}
\keyword{filtration,}

