% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_wat.r
\name{dens_wat}
\alias{dens_wat}
\title{dens_wat Function
This function calculates the density of water at the specified temperature (°C). The density units are g / cm3.}
\usage{
dens_wat(TEMP = 25)
}
\arguments{
\item{TEMP}{Temperature in units of °C. Defaults to a value of 25°C}
}
\description{
dens_wat Function
This function calculates the density of water at the specified temperature (°C). The density units are g / cm3.
}
\examples{
dens_wat()
[1] 0.99708
}
\keyword{Density}
\keyword{Water}
\keyword{filtration,}
\keyword{of}

