% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_filt.r
\name{L_filt}
\alias{L_filt}
\title{L_filt Function
This function calculates the distance in centimeters to reduce the suspended concentration by a
factor of C_C0 (i.e., C/C0).}
\usage{
L_filt(C_C0 = 1e-06, alpha = 1, Kx = 0.02, GRAD = 7e-04, POR = 0.21,
  K_units = "cm/s", TEMP = 25, DIA_SUSP = 1e-04, DIA_SOIL = 0.053,
  PART_DENS = 2.64)
}
\arguments{
\item{alpha}{Factor that accounts for chemical affects on particle removal. Defaults to a value of 1 (i.e., 
chemical affects do not enhance particle removal}

\item{Kx}{Defaults to a value of 0.02 cm/s}

\item{GRAD}{Defaults to a value of 0.0007}

\item{POR}{Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}

\item{K_units}{Units of hydraulic conductivity. Defaults to cm/s but also accepts ft/day. Units should be surrounded
by quotation marks to ensure it is treated as a string of text.}

\item{TEMP}{Temperature in units of °C. Defaults to a value of 25°C}

\item{DIA_SUSP}{Diameter of the suspended solids in units of cm}

\item{DIA_SOIL}{Diameter of the soil particles in units of cm}

\item{PART_DENS}{Density of the suspended particle in units of g/cm³}

\item{C_CO}{Target TSS concentration reduction. Defaults to a value of 1e-6. 
That is to say that the TSS reduction is a factor of 1,000,000 times}
}
\description{
L_filt Function
This function calculates the distance in centimeters to reduce the suspended concentration by a
factor of C_C0 (i.e., C/C0).
}
\examples{
L_filt()
[1] 0.2625636
}
\keyword{Length}
\keyword{filtration,}

