% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ng_parm.r
\name{Ng_parm}
\alias{Ng_parm}
\title{Ng_parm Function
This function calculates the ratio of the Stokes' settling velocity of the particle to the superficial filtration velocity}
\usage{
Ng_parm(Kx = 0.02, GRAD = 7e-04, POR = 0.21, K_units = "cm/s",
  TEMP = 25, DIA_SUSP = 1e-04, DIA_SOIL = 0.053, PART_DENS = 2.64)
}
\arguments{
\item{Kx}{Defaults to a value of 0.02 cm/s}

\item{GRAD}{Defaults to a value of 0.0007}

\item{POR}{Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}

\item{K_units}{Units of hydraulic conductivity. Defaults to cm/s but also accepts ft/day. Units should be surrounded
by quotation marks to ensure it is treated as a string of text.}

\item{TEMP}{Temperature in units of °C. Defaults to a value of 25°C}

\item{DIA_SUSP}{Diameter of the suspended solids in units of cm}

\item{DIA_SOIL}{Diameter of the soil particles in units of cm}

\item{PART_DENS}{Density of the suspended particle in units of g/cm³}
}
\description{
Ng_parm Function
This function calculates the ratio of the Stokes' settling velocity of the particle to the superficial filtration velocity
}
\examples{
Ng_parm()
[1] 0.2422003
}
\keyword{Ng}
\keyword{filtration,}

