% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLo_parm.r
\name{NLo_parm}
\alias{NLo_parm}
\title{NLo_parm Function
This function calculates the dimensionless London Van-Der Walls force number.}
\usage{
NLo_parm(Kx = 0.02, GRAD = 7e-04, POR = 0.21, K_units = "cm/s",
  TEMP = 25, DIA_SUSP = 1e-04)
}
\arguments{
\item{Kx}{Defaults to a value of 0.02 cm/s}

\item{GRAD}{Defaults to a value of 0.0007}

\item{POR}{Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}

\item{K_units}{Units of hydraulic conductivity. Defaults to cm/s but also accepts ft/day. Units should be surrounded
by quotation marks to ensure it is treated as a string of text.}

\item{TEMP}{Temperature in units of °C. Defaults to a value of 25°C}

\item{DIA_SUSP}{Diameter of the suspended solids in units of cm}
}
\description{
NLo_parm Function
This function calculates the dimensionless London Van-Der Walls force number.
}
\examples{
 NLo_parm()
[1] 0.0002383168
}
\keyword{Van-Der}
\keyword{Walls}
\keyword{filtration,}

