% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Npe_parm.r
\name{Npe_parm}
\alias{Npe_parm}
\title{Npe_parm Function
This function calculates the dimensionless Peclet number.}
\usage{
Npe_parm(Kx = 0.02, GRAD = 7e-04, POR = 0.21, K_units = "cm/s",
  TEMP = 25, DIA_SUSP = 1e-04, DIA_SOIL = 0.053)
}
\arguments{
\item{Kx}{Defaults to a value of 0.02 cm/s}

\item{GRAD}{Defaults to a value of 0.0007}

\item{POR}{Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}

\item{K_units}{Units of hydraulic conductivity. Defaults to cm/s but also accepts ft/day. Units should be surrounded
by quotation marks to ensure it is treated as a string of text.}

\item{TEMP}{Temperature in units of °C. Defaults to a value of 25°C}

\item{DIA_SUSP}{Diameter of the suspended solids in units of cm}

\item{DIA_SOIL}{Diameter of the soil particles in units of cm}
}
\description{
Npe_parm Function
This function calculates the dimensionless Peclet number.
}
\examples{
Npe_parm(DIA_SOIL = c(0.001, 0.01, 0.1, 1))
[1]   0.5436584   5.4365837  54.3658370 543.6583699
}
\keyword{filtration,}
\keyword{peclet}

