% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nr_parm.r
\name{Nr_parm}
\alias{Nr_parm}
\title{Nr_parm Function
This function calculates the ratio of the suspended particle to soil particle size ratio}
\usage{
Nr_parm(DIA_SUSP = 1e-04, DIA_SOIL = 0.053)
}
\arguments{
\item{DIA_SUSP}{Diameter of the suspended solids in units of cm}

\item{DIA_SOIL}{Diameter of the soil particles in units of cm}
}
\description{
Nr_parm Function
This function calculates the ratio of the suspended particle to soil particle size ratio
}
\examples{
Nr_parm()
[1] 1.886792e-05
}
\keyword{Ratio}
\keyword{filtration,}
\keyword{of}
\keyword{particle}
\keyword{sizes}

