% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_parm.r
\name{p_parm}
\alias{p_parm}
\title{p_parm Function
This function calculates the p-parameter from Table 1 of Tobiason and O'Melia (1988). The p-parameter is based on soil porosity.}
\usage{
p_parm(POR = 0.21)
}
\arguments{
\item{POR}{porosity. Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}
}
\description{
p_parm Function
This function calculates the p-parameter from Table 1 of Tobiason and O'Melia (1988). The p-parameter is based on soil porosity.
}
\examples{
p_parm(POR = 0.3)
[1] 0.887904

p_parm(POR = 21)
[1] "Error: Porosity value must be less than 1"
}
\keyword{filtration,}
\keyword{p-parameter}
\keyword{porosity,}

