% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_parm.r
\name{U_parm}
\alias{U_parm}
\title{U_parm Function to calculate groundwater velocity
This function calculates the U-parameter grounwdater velocity from Table 1 of Tobiason and O'Melia (1988). 
The U-parameter is based on the Kx (hydraulic conductivity), GRAD (hydraulic gradient), and POR (porosity).
The default units are cm/s but can be changed with K_units. Only ft/day are accepted as alternative units.
The units of the output are cm/s.}
\usage{
U_parm(Kx = 0.02, GRAD = 7e-04, POR = 0.21, K_units = "cm/s")
}
\arguments{
\item{Kx}{Hydraulic conductivity. Defaults to a value of 0.02 cm/s}

\item{GRAD}{Hydraulic gradient. Defaults to a value of 0.0007}

\item{POR}{porosity. Defaults to a value of 0.21. The value of 0.21 is the average for medium sand as reported by Johnson (1967)}

\item{K_units}{Units of hydraulic conductivity. Defaults to cm/s but also accepts ft/day. Units should be surrounded
by quotation marks to ensure it is treated as a string of text.}
}
\description{
U_parm Function to calculate groundwater velocity
This function calculates the U-parameter grounwdater velocity from Table 1 of Tobiason and O'Melia (1988). 
The U-parameter is based on the Kx (hydraulic conductivity), GRAD (hydraulic gradient), and POR (porosity).
The default units are cm/s but can be changed with K_units. Only ft/day are accepted as alternative units.
The units of the output are cm/s.
}
\examples{
U <- U_parm(Kx = 0.02, GRAD = 0.0007, POR = 0.21, K_units = "cm/s")
U
[1] 6.666667e-05
}
\keyword{U-parameter}
\keyword{filtration,}
\keyword{groundwater}
\keyword{velocity,}

