% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visc_wat.r
\name{visc_wat}
\alias{visc_wat}
\title{visc_wat Function
This function calculates the viscosity of water for a given temperature (°C). The units of viscosity are g/cm/s}
\usage{
visc_wat(TEMP = 25)
}
\arguments{
\item{TEMP}{porosity. Defaults to a value of 25°C.}
}
\description{
visc_wat Function
This function calculates the viscosity of water for a given temperature (°C). The units of viscosity are g/cm/s
}
\examples{
visc_wat(TEMP = 25)
[1] 0.0089044
}
\keyword{filtration,}
\keyword{temperature,}
\keyword{viscosity}

